+++ service/Models/CustomerHandler.go
@@ -19,3 +19,10 @@ func ValidateCustomer(customer *Customer, customerID uint) (err error) {
 	}
 	return nil
 }
+
+func GetLastOrderTime(customer Customer, customerID uint) (err error) {
+	if err := Config.DB.First(&customer, customerID).Error; err != nil {
+		return err
+	}
+	return nil
+}
